requirements.txt
psycopg2-binary==2.9.3
requests==2.26.0
pytest==7.0.1